pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "talimini"
version = "0.1.0"
description = "Flesh and Blood card game engine with reinforcement learning capabilities"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Talimini Contributors"},
]
keywords = ["card-game", "flesh-and-blood", "reinforcement-learning", "game-engine", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies
dependencies = [
    "numpy>=1.20.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
# Machine learning dependencies
ml = [
    "torch>=2.0.0",
    "stable-baselines3>=2.0.0",
    "sb3-contrib>=2.0.0",
    "gymnasium>=0.28.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# All dependencies
all = [
    "talimini[ml,dev]",
]

[project.scripts]
fabgame = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/talimini"
Repository = "https://github.com/yourusername/talimini"
Issues = "https://github.com/yourusername/talimini/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fabgame*", "scripts*", "tools*"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 120
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we'll fix this during refactoring)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually during refactoring
disallow_any_generics = false
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yaml.*",
    "torch.*",
    "stable_baselines3.*",
    "sb3_contrib.*",
    "gymnasium.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=fabgame",
    "--cov-report=term-missing",
    "--cov-report=html",
]
